# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "BedOut": {
        "properties": {
          "aspect_deg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aspect Deg"
          },
          "bed_id": {
            "title": "Bed Id",
            "type": "integer"
          },
          "constraint_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Constraint Hint"
          },
          "ellenberg_l": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ellenberg L"
          },
          "ellenberg_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ellenberg M"
          },
          "ellenberg_n": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ellenberg N"
          },
          "ellenberg_r": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ellenberg R"
          },
          "height_above_ground": {
            "title": "Height Above Ground",
            "type": "number"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "light_computed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Light Computed At"
          },
          "moisture": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Moisture"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "plantings": {
            "items": {
              "$ref": "#/components/schemas/PlantingOut"
            },
            "title": "Plantings",
            "type": "array"
          },
          "points": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "type": "number"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Points"
          },
          "polygon": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "title": "Polygon",
            "type": "array"
          },
          "shape": {
            "title": "Shape",
            "type": "string"
          },
          "slope_deg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slope Deg"
          },
          "soil_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Soil Type"
          },
          "sun_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sun Hours"
          },
          "width": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "x": {
            "title": "X",
            "type": "number"
          },
          "y": {
            "title": "Y",
            "type": "number"
          }
        },
        "required": [
          "bed_id",
          "name",
          "polygon",
          "kind",
          "shape",
          "x",
          "y",
          "points",
          "width",
          "constraint_hint",
          "soil_type",
          "moisture",
          "ellenberg_l",
          "ellenberg_m",
          "ellenberg_n",
          "ellenberg_r",
          "sun_hours",
          "slope_deg",
          "aspect_deg",
          "light_computed_at",
          "height_above_ground",
          "label",
          "plantings"
        ],
        "title": "BedOut",
        "type": "object"
      },
...
      "GardenOut": {
        "properties": {
          "beds": {
            "items": {
              "$ref": "#/components/schemas/BedOut"
            },
            "title": "Beds",
            "type": "array"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "moisture": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Moisture"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "observed_colours": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Observed Colours",
            "type": "object"
          },
          "obstacles": {
            "items": {
              "$ref": "#/components/schemas/ObstacleOut"
            },
            "title": "Obstacles",
            "type": "array"
          },
          "share_token": {
            "title": "Share Token",
            "type": "string"
          },
          "soil_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Soil Type"
          },
          "unidentified_plantings": {
            "default": 0,
            "title": "Unidentified Plantings",
            "type": "integer"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "share_token",
          "name",
          "latitude",
          "longitude",
          "created_at",
          "updated_at",
          "soil_type",
          "moisture",
          "observed_colours",
          "beds",
          "obstacles"
        ],
        "title": "GardenOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HeightReport": {
        "description": "Where the heights around this garden came from.\n\nReported because an assumed height presented as a measured one is the same\nlie as a filter that hides what it dropped.",
        "properties": {
          "assumed": {
            "title": "Assumed",
            "type": "integer"
          },
          "estimated": {
            "title": "Estimated",
            "type": "integer"
          },
          "measured": {
            "title": "Measured",
            "type": "integer"
          }
        },
        "required": [
          "measured",
          "estimated",
          "assumed"
        ],
        "title": "HeightReport",
        "type": "object"
      },
      "ImageryOut": {
        "description": "The aerial imagery available at a place, if any.\n\n`attribution` is a condition of the licence, not a caption \u2014 DL-DE/BY-2.0\nand CC-BY-4.0 both require the named credit, so imagery shown without it is\nimagery used outside its terms.",
        "properties": {
          "attribution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attribution"
          },
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "layer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Layer"
          },
          "licence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Licence"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "available"
        ],
        "title": "ImageryOut",
        "type": "object"
      },
...
      "LatLonIn": {
        "properties": {
          "lat": {
            "maximum": 55.2,
            "minimum": 47.0,
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "maximum": 15.2,
            "minimum": 5.5,
            "title": "Lon",
            "type": "number"
          }
        },
        "required": [
          "lat",
          "lon"
        ],
        "title": "LatLonIn",
        "type": "object"
      },
...
      "MapGardenOut": {
        "description": "A garden created from the map, and what the map could and could not say.",
        "properties": {
          "garden": {
            "$ref": "#/components/schemas/GardenOut"
          },
          "heights": {
            "$ref": "#/components/schemas/HeightReport"
          }
        },
        "required": [
          "garden",
          "heights"
        ],
        "title": "MapGardenOut",
        "type": "object"
      },
      "MapSelection": {
        "properties": {
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "neighbourhood": {
            "default": "detached",
            "title": "Neighbourhood",
            "type": "string"
          },
          "outline": {
            "items": {
              "$ref": "#/components/schemas/LatLonIn"
            },
            "maxItems": 500,
            "minItems": 3,
            "title": "Outline",
            "type": "array"
          }
        },
        "required": [
          "name",
          "outline"
        ],
        "title": "MapSelection",
        "type": "object"
      },
...
      "ObstacleOut": {
        "properties": {
          "constraint_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Constraint Hint"
          },
          "eaves_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eaves M"
          },
          "eaves_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eaves Source"
          },
          "footprint": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "title": "Footprint",
            "type": "array"
          },
          "height": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "height_source": {
            "title": "Height Source",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "obstacle_id": {
            "title": "Obstacle Id",
            "type": "integer"
          },
          "outline_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outline Source"
          },
          "points": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "type": "number"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Points"
          },
          "roof": {
            "title": "Roof",
            "type": "string"
          },
          "roof_fall_deg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roof Fall Deg"
          },
          "roof_lines": {
            "items": {
              "items": {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              "type": "array"
            },
            "title": "Roof Lines",
            "type": "array"
          },
          "roof_pitch_deg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roof Pitch Deg"
          },
          "roof_source": {
            "title": "Roof Source",
            "type": "string"
          },
          "shadow": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shadow"
          },
          "shape": {
            "title": "Shape",
            "type": "string"
          },
          "width": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "x": {
            "title": "X",
            "type": "number"
          },
          "y": {
            "title": "Y",
            "type": "number"
          }
        },
        "required": [
          "obstacle_id",
          "kind",
          "roof",
          "eaves_m",
          "roof_source",
          "eaves_source",
          "roof_fall_deg",
          "roof_pitch_deg",
          "roof_lines",
          "label",
          "height_source",
          "x",
          "y",
          "shape",
          "width",
          "points",
          "constraint_hint",
          "height",
          "footprint"
        ],
        "title": "ObstacleOut",
        "type": "object"
      },
...
      "PlaceOut": {
        "properties": {
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "title": "Lon",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "lat",
          "lon"
        ],
        "title": "PlaceOut",
        "type": "object"
      },
      "PlaceSearchOut": {
        "properties": {
          "places": {
            "items": {
              "$ref": "#/components/schemas/PlaceOut"
            },
            "title": "Places",
            "type": "array"
          }
        },
        "required": [
          "places"
        ],
        "title": "PlaceSearchOut",
        "type": "object"
      },
...
      "PlantingOut": {
        "properties": {
          "added_at": {
            "title": "Added At",
            "type": "string"
          },
          "canonical_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Name"
          },
          "planting_id": {
            "title": "Planting Id",
            "type": "integer"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "raw_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Name"
          },
          "taxon_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taxon Id"
          },
          "x": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "X"
          },
          "y": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Y"
          }
        },
        "required": [
          "planting_id",
          "taxon_id",
          "canonical_name",
          "raw_name",
          "quantity",
          "added_at",
          "x",
          "y"
        ],
        "title": "PlantingOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "NinaNatur",
    "version": "contract"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/gardens/from-map": {
      "post": {
        "description": "Create a garden from an outline drawn on the map, with what shades it.\n\nThe margin is 50 m and objects are filtered by whether their shadow could\narrive at all \u2014 generous where it matters, quiet where it does not.",
        "operationId": "garden_from_map_api_v1_gardens_from_map_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MapSelection"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MapGardenOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Garden From Map",
        "tags": [
          "geo"
        ]
      }
    },
...
    "/api/v1/geo/imagery": {
      "get": {
        "description": "Which state's orthophotos may be shown here, if any.\n\nPer Bundesland because the licences are: there is no federal source, and a\nstate without an entry gets no imagery rather than a neighbour's.",
        "operationId": "imagery_at_api_v1_geo_imagery_get",
        "parameters": [
          {
            "in": "query",
            "name": "lat",
            "required": true,
            "schema": {
              "maximum": 55.2,
              "minimum": 47.0,
              "title": "Lat",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "lon",
            "required": true,
            "schema": {
              "maximum": 15.2,
              "minimum": 5.5,
              "title": "Lon",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Imagery At",
        "tags": [
          "geo"
        ]
      }
    },
    "/api/v1/geo/search": {
      "get": {
        "description": "Addresses matching a query. Empty in, empty out \u2014 a keystroke-per-request\nsearch would be a poor way to treat a free community service.",
        "operationId": "find_place_api_v1_geo_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 200,
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceSearchOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Find Place",
        "tags": [
          "geo"
        ]
      }
    },
...
  }
}